plot.variogram.SpATS {SpATS}	R Documentation

Default variogram.SpATS plotting

Description

Takes a fitted variogram.SpATS object produced by variogram.SpATS() and plots the associated sample variogram using an RGL 3D perspective plot (package plot3Drgl).

Usage

## S3 method for class 'variogram.SpATS'
plot(x, min.length = 30, ...)

Arguments

x
an object of class variogram.SpATS as produced by variogram.SpATS().

min.length
numerical value. The sample variogram is depicted including only those pairs with more than min.length observations (see variogram.SpATS).

...
further arguments passed to or from other methods. Not yet implemented.

Details

This function as well as function variogram.SpATS() can only be used for regular two dimensional data.

References

Gilmour, A.R., Cullis, B.R., and Verbyla, A.P. (1997). Accounting for Natural and Extraneous Variation in the Analysis of Field Experiments. Journal of Agricultural, Biological, and Environmental Statistics, 2, 269 - 293.

Stefanova, K.T., Smith, A.B., and Cullis, B.R. (2009). Enhanced Diagnostics for the Spatial Analysis of Field Trials. Journal of Agricultural, Biological, and Environmental Statistics, 14, 392 - 410.

See Also

SpATS, variogram.SpATS

Examples

library(SpATS)
data(wheatdata)
wheatdata$R <- as.factor(wheatdata$row)
wheatdata$C <- as.factor(wheatdata$col)

m0 <- SpATS(response = "yield", spatial = ~ SAP(col, row, nseg = c(10,20), degree = 3, pord = 2), 
 genotype = "geno", fixed = ~ colcode + rowcode, random = ~ R + C, data = wheatdata, 
 control =  list(tolerance = 1e-03))

# Compute the variogram
var.m0 <- variogram(m0)
# Plot the variogram
plot(var.m0)

[Package SpATS version 1.0-19 Index]