print.SpATS {SpATS}	R Documentation

Print method for SpATS objects

Description

Default print method for objects fitted with SpATS() function.

Usage

## S3 method for class 'SpATS'
print(x, ...)

Arguments

x
an object of class SpATS as produced by SpATS()

...
further arguments passed to or from other methods. Not yet implemented.

Details

A short summary is printed including: the variable under study (response), the variable containing the genotypes, the spatial model, and the random and fixed components (when appropriate). Besides this information, the number of observations used to fit the model, as well as of those deleted due to missingness or zero weights are reported. Finally, the effective degrees of freedom (effective dimension) of the fitted model and the (REML) deviance is also printed.

See Also

SpATS, summary.SpATS

Examples

library(SpATS)
data(wheatdata)
wheatdata$R <- as.factor(wheatdata$row)
wheatdata$C <- as.factor(wheatdata$col)

m0 <- SpATS(response = "yield", spatial = ~ SAP(col, row, nseg = c(10,20), degree = 3, pord = 2), 
 genotype = "geno", fixed = ~ colcode + rowcode, random = ~ R + C, data = wheatdata, 
 control =  list(tolerance = 1e-03))

m0

[Package SpATS version 1.0-19 Index]